Sensor.ino
When turn Arduino on you can input a value which you want the water level to be at, which will give you a return of what the desired reading is-sometimes when you write a 2 digit number it only holds the last digit e.g. 45 can return as 5 but if you type it again a few times it should lock in. Then key in  'on' in serial monitor and the pumps should react accordingly, and automatically turn off when reaches the desired distance. And can key in 'q' to turn the pumps off.